Cambios del main3 al main2 también actualicé la lista del cmake con el ejecutable main2
